+++ CHANGELOG.md
@@ -9,7 +9,8 @@
 - `@dicebear/avatars` New `rotate` option.
 - `@dicebear/avatars` New `flip` option.
 - `@dicebear/avatars` New `size` option as replacement for `width` and `height` options.
-- `@dicebear/avatars` New `x` any `y` translate options.
+- `@dicebear/avatars` New `translateX` option.
+- `@dicebear/avatars` New `translateY` option.
 
 ### Changed